requirements.txt
Django>=5.0,<6.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-allauth
python-decouple
psycopg2-binary
dj-database-url
Pillow
django-storages
boto3
drf-spectacular
django-filter
celery
redis
django-celery-beat
django-push-notifications
twilio
sendgrid
django-debug-toolbar
django-extensions
faker
pytest
pytest-django
factory-boy
gunicorn
whitenoise
